V2 (#7)
Добавлена возможность получения бумажной открытки, мобильная версия приложения, а также улучшена визуальная составляющая проекта в целом.

+++ utils.py
@@ -4,7 +4,7 @@ import qrcode
 from hashlib import sha256
 import time
 QR_DIR = "static/qrcodes"
-EXPIRATION_TIME = 1
+EXPIRATION_TIME = 5*60
 
 os.makedirs(QR_DIR, exist_ok=True)
 
@@ -44,6 +44,30 @@ def delete_qrcodes():
     
     for filename in os.listdir(QR_DIR):
         file_path = os.path.join(QR_DIR, filename)
+        if os.path.isfile(file_path) and current_time - os.path.getctime(file_path) > EXPIRATION_TIME:
+            try:
+                os.remove(file_path)
+                deleted_files_count += 1
+                print(f"Удален файл: {file_path}")
+            except Exception as e:
+                print(f"Ошибка при удалении файла {file_path}: {e}")
+
+    return deleted_files_count
+
+def delete_postcards():
+    """
+    Удаляет старые файлы открыток из директории static/postcards.
+    Возвращает количество удаленных файлов.
+    """
+    postcards_dir = os.path.join("static", "postcards")
+    if not os.path.exists(postcards_dir):
+        raise FileNotFoundError(f"Папка с открытками не найдена: {postcards_dir}")
+    
+    current_time = time.time()
+    deleted_files_count = 0
+    
+    for filename in os.listdir(postcards_dir):
+        file_path = os.path.join(postcards_dir, filename)
         if os.path.isfile(file_path) and current_time - os.path.getctime(file_path) > EXPIRATION_TIME:
             try:
                 os.remove(file_path)
